Use sources generated by kapt rather than classfiles
This fixes problems with the incremental compilation as well as
with the incremental build with kapt generating stubs.

Previously java output directory was added to kotlin-after-java
classpath, so kotlin code could refer a generated code.

This way the following scenario would've taken place:
// first build, kotlin-after-java
1. gradle takes input files snapshot (including classes in
java destination dir since they are in task's classpath)
2. kotlin code compiles and classes are copied to java destination dir
// second build
3. gradle detects that input files differ from snapshots taken on step 1

Because of that kotlin-after-java task would be executed
even if kotlin code was not changed and incremental compilation
would not work at all, since we treat any change in class-file
as a trigger to rebuild.

The commit removes java output dir from kotlin-after-java classpath
and adds generated sources folder as a source root instead.
As a side-effect kotlin is not able to refer to refer to
generated java sources that do not follow package-class
layout convention (i.e to class A in file B.java).
This behavior is consistent with normal java source files (see KT-4455).